move epicIgnoreAction into mwp-api-state
