+++ packages/mwp-api-state/src/sync/index.test.js
@@ -3,7 +3,8 @@ import 'rxjs/add/operator/toPromise';
 import { ActionsObservable } from 'redux-observable';
 import { CLICK_TRACK_CLEAR_ACTION } from 'mwp-tracking-plugin/lib/util/clickState'; // mwp-tracking-plugin/util/clickState
 import { LOCATION_CHANGE, SERVER_RENDER } from 'mwp-router'; // mwp-router
-import { createFakeStore, epicIgnoreAction } from 'mwp-test-utils';
+import { createFakeStore } from 'mwp-test-utils';
+import { epicIgnoreAction } from '../util/testUtils';
 
 import fetch from 'node-fetch';
 global.fetch = fetch;
